// backend/src/auth/dto/register.dto.ts
export class RegisterDto {
  firstName: string = '';
  lastName: string = '';
  email: string = '';
  password: string = '';
  confirmPassword: string = '';
  phoneNumber: string = '';
  expertise: string  = '';
  bio: string = '';
  education: string = '';
  experience: string; 
  teachingAreas: string[];
}